# internal/docs/swagger.yaml
definitions:
...
  handler.createPatientRequest:
    properties:
      allergies:
        type: string
      blood_group:
        type: string
      current_medication:
        type: string
      date_of_birth:
        type: string
      emergency_contact:
        type: string
      gender:
        type: string
      medical_history:
        type: string
    required:
    - date_of_birth
    - gender
    type: object
...
  handler.patientResponse:
    properties:
      allergies:
        type: string
      blood_group:
        type: string
      current_medication:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      emergency_contact:
        type: string
      gender:
        type: string
      id:
        type: integer
      medical_history:
        type: string
      name:
        type: string
      user_id:
        type: integer
    type: object
...
  handler.updatePatientRequest:
    properties:
      allergies:
        type: string
      blood_group:
        type: string
      current_medication:
        type: string
      date_of_birth:
        type: string
      emergency_contact:
        type: string
      gender:
        type: string
      medical_history:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /patients:
    post:
      consumes:
      - application/json
      description: Create a new patient profile
      parameters:
      - description: Patient Information
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/handler.createPatientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created patient profile
          schema:
            $ref: '#/definitions/handler.patientResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create patient profile
      tags:
      - patients
  /patients/{id}:
    delete:
      description: Delete a patient profile by ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete patient profile
      tags:
      - patients
    get:
      description: Get a patient profile by ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Patient profile
          schema:
            $ref: '#/definitions/handler.patientResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get patient profile
      tags:
      - patients
    put:
      consumes:
      - application/json
      description: Update an existing patient profile
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patient Information
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/handler.updatePatientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated patient profile
          schema:
            $ref: '#/definitions/handler.patientResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update patient profile
      tags:
      - patients
...
  /patients/user/{userId}:
    get:
      description: Get a patient profile by user ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Patient profile
          schema:
            $ref: '#/definitions/handler.patientResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get patient profile by user ID
      tags:
      - patients
...
swagger: "2.0"